LBHttpSolrServer allows to specify a preferred server (previously obtained) for a query.
If the preferred server is no longer alive the request is retried on another live server.
After live servers are exhausted, any servers previously marked as dead will be tried before
failing the request. If no live servers are found a SolrServerException is thrown.

The existing request functionality is slightly refactored to allow code reuse for this
new feature (reduce code duplication).